requirements.txt
nudenet
Pillow
transformers
torch --index-url https://download.pytorch.org/whl/cpu
